handle delete Tag and Category

+++ app/Http/Controllers/CategoriesController.php
@@ -87,6 +87,9 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if($category->posts()->count() > 0){
+            return redirect()->back()->with('error', 'Category can\'t be deleted beacuse it has some posts.');
+        }
         $category->delete();
         return redirect()->route('categories.index')
            ->with('success', 'deleted successfluy');
